# PERF: ArrowExtensionArray.to_numpy