- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.

There are a few different cases/paths in `ArrowExtensionArray.to_numpy` where copying can be avoided and more efficient fillna logic can be used:

```
import pandas as pd
import numpy as np

N = 10_000_000
data = np.random.randn(N)


# -----------------------------------------------------------------------------------
# case 1: no nulls

arr = pd.array(data, dtype="float64[pyarrow]")

%timeit arr.to_numpy("float64")

# 16 ms ± 620 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)        -> main
# 19.8 µs ± 711 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  -> PR


# -----------------------------------------------------------------------------------
# case 2: contains nulls

arr = pd.array(data, dtype="float64[pyarrow]")
arr[::1000] = None

%timeit arr.to_numpy("float64", na_value=np.nan)

# 75.3 ms ± 3.48 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> main
# 17 ms ± 204 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)    -> PR


# -----------------------------------------------------------------------------------
# case 3: null pyarrow type

arr = pd.array([None] * N, dtype="null[pyarrow]")

%timeit arr.to_numpy("float64", na_value=np.nan)

# 324 ms ± 14.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)     -> main
# 8.58 ms ± 214 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  -> PR
```